Added `[-i|--integrity] {IntegrityLevel}` to specify a desired Integrity Level. gsudo can now elevate or restrict the permissions of a given command. Added `status` verb to list info on the current console/gsudo session.

+++ src/gsudo/Commands/CtrlCCommand.cs
@@ -15,7 +15,7 @@ namespace gsudo.Commands
         {            
             FreeConsole();
 
-            if (AttachConsole((uint)Pid))
+            if (AttachConsole(Pid))
             {
                 if (sendSigBreak)
                     GenerateConsoleCtrlEvent(CtrlTypes.CTRL_BREAK_EVENT, 0);
